% Generated by roxygen2 (4.0.1): do not edit by hand
\name{appendSqlCcdData}
\alias{appendSqlCcdData}
\title{appendSqlCcdData}
\usage{
appendSqlCcdData(object, oStratumId, oRowId, oY, oTime, cRowId, cCovariateId,
  cCovariateValue)
}
\description{
\code{appendSqlCcdData} appends data to an OHDSI data object.
}

