% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{cacheCyclopsModelForJava}
\alias{cacheCyclopsModelForJava}
\title{Place Cyclops model fit object in a persistent cache}
\usage{
cacheCyclopsModelForJava(object)
}
\arguments{
\item{object}{Cyclops model fit object}
}
\value{
Integer index of object in persistent cache
}
\description{
\code{cacheCyclopsModelForJava} places a Cyclops model fit object into a persistent cache,
so that the object may be used across language-barriers.
}
