% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ccd_model_data}
\alias{ccd_model_data}
\title{ccd_model_data}
\usage{
ccd_model_data(pid, y, z, offs, dx, sx, ix)
}
\arguments{
\item{pid}{Vector of patient identifiers (function assumes these are sorted)}

\item{y}{Vector of outcomes}

\item{z}{Vector of secondary outcomes (or NULL if unneeded for model)}

\item{offs}{Vector of regression model offsets (or NULL)}

\item{dx}{Dense matrix of covariates (or NULL)}

\item{sx}{Sparse matrix of covariates (or NULL)}

\item{ix}{Indicator matrix of covariates (or NULL)}
}
\value{
A CCD model data object
}
\description{
\code{ccd_model_data} creates a CCD model data object
}
\details{
This function is fun.  This function currently creates a deep copy of all data.
Another deep copy is also then made during CCD engine initialization; one of
these copies should be removed.
}
\examples{
splitSql("SELECT * INTO a FROM b; USE x; DROP TABLE c;")
}

