% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{clearCyclopsModelCache}
\alias{clearCyclopsModelCache}
\title{Clear the Cyclops persistent cache}
\usage{
clearCyclopsModelCache()
}
\description{
\code{clearCyclopsModelCache} clears the persistent cache holding
Cyclops model fit objects that may be used across language-barriers.
}
