% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coef.ccdFit}
\alias{coef.ccdFit}
\title{Extract model coefficients}
\usage{
\method{coef}{ccdFit}(object, ...)
}
\arguments{
\item{object}{OHDSI CCD model fit object}

\item{...}{Other arguments}
}
\value{
Named numeric vector of model coefficients.
}
\description{
\code{coef.ccdFit} extracts model coefficients from an OHDSI CCD model fit object
}

