% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coef.cyclopsFit}
\alias{coef.cyclopsFit}
\title{Extract model coefficients}
\usage{
\method{coef}{cyclopsFit}(object, ...)
}
\arguments{
\item{object}{Cyclops model fit object}

\item{...}{Other arguments}
}
\value{
Named numeric vector of model coefficients.
}
\description{
\code{coef.cyclopsFit} extracts model coefficients from an Cyclops model fit object
}

