% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cyclopsCoxMM.R
\name{compareCoxMM}
\alias{compareCoxMM}
\title{Simulates data to test cox MM}
\usage{
compareCoxMM(x, y, e, intercept = TRUE)
}
\arguments{
\item{x}{covariates}

\item{y}{outcomes}

\item{e}{distance between Poisson regressions to stop}

\item{intercept}{whether to include intercept in model, default = TRUE}
}
\value{
Returns a list with the following: \describe{
\item{coefficients}{calculated coefficients for each method}
\item{cyclopsCox}{iterations ran for each method}
\item{goldPoisson}{distance between the two cox methods, two poisson methods, and between the two cyclops methods} }
}
\description{
Compares gold standard and cyclops calculations of original cox regression and poisson MM regression
}

