% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cyclopsCoxMM.R
\name{computeCoxMM}
\alias{computeCoxMM}
\title{computes Cox MM}
\usage{
computeCoxMM(x, y, e, intercept, method)
}
\arguments{
\item{x}{covariate data}

\item{y}{outcome data}

\item{e}{stopping criteria - norm of beta}

\item{intercept}{whether to include intercept}

\item{method}{"glm" or "cyclops"}
}
\value{
\describe{
\item{coefficients}{calculated coefficients}
\item{iterations}{total iterations ran}
\item{runs}{number of poisson regressions ran}
\item{iterCounts}{iterations per poisson regression}}
}
\description{
computes the Cox MM poisson regressions
}

