% Generated by roxygen2 (4.0.2): do not edit by hand
\name{confint.cyclopsFit}
\alias{confint.cyclopsFit}
\title{confint.cyclopsFit}
\usage{
\method{confint}{cyclopsFit}(object, parm, level = 0.95, control,
  overrideNoRegularization = FALSE, includePenalty = FALSE, ...)
}
\arguments{
\item{object}{A fitted Cyclops model object}

\item{parm}{A specification of which parameters require confidence intervals,
either a vector of numbers of covariateId names}

\item{level}{Numeric: confidence level required}

\item{control}{A Cyclops \code{\link{control}} object}

\item{overrideNoRegularization}{Logical: Enable confidence interval estimation for regularized parameters}

\item{includePenalty}{Logical: Include regularized covariate penalty in profile}

\item{...}{Additional argument(s) for methods}
}
\value{
A matrix with columns reporting lower and upper confidence limits for each parameter.
These columns are labelled as (1-level) / 2 and 1 - (1 - level) / 2 in %
(by default 2.5% and 97.5%)
}
\description{
\code{confinit.cyclopsFit} profiles the data likelihood to construct confidence intervals of
arbitrary level.   TODO: Profile data likelihood or joint distribution of remaining parameters.
}

