% Generated by roxygen2 (4.0.1): do not edit by hand
\name{control}
\alias{control}
\title{control}
\usage{
control(maxIterations = 1000, tolerance = 1e-08,
  convergenceType = "gradient")
}
\arguments{
\item{maxIterations}{Integer: maximum iterations of CCD to attempt before returning a failed-to-converge error}

\item{tolerance}{Numeric: maximum change in convergence criterion from successive iterations to achieve convergence}

\item{convergenceType}{String: name of convergence criterion to employ (described in more detail below)}
}
\value{
A CCD convergence criteria object of class inheriting from \code{"ccdConvergence"} for use with \code{fitCcdModel}.
}
\description{
\code{control} builds a CCD convergence criteria object
#'
}
\section{Criteria}{

}

