% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convertToCcdVariance}
\alias{convertToCcdVariance}
\title{Convert to CCD Prior Variance}
\usage{
convertToCcdVariance(lambda, nobs)
}
\arguments{
\item{lambda}{Regularization parameter from \code{glmnet}}

\item{nobs}{Number of observation rows in dataset}
}
\value{
Prior variance under a Laplace() prior
}
\description{
\code{convertToCcdVariance} converts the regularization parameter \code{lambda}
from \code{glmnet} into a prior variance.
}

