% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertToGlmnetLambda}
\alias{convertToGlmnetLambda}
\title{Convert to glmnet regularization parameter}
\usage{
convertToGlmnetLambda(variance, nobs)
}
\arguments{
\item{variance}{Prior variance}

\item{nobs}{Number of observation rows in dataset}
}
\value{
\code{lambda}
}
\description{
\code{convertToGlmnetLambda} converts a prior variance
from \code{Cyclops} into the regularization parameter \code{lambda}.
}

