% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeEffects.R
\name{convertToLongOutcome}
\alias{convertToLongOutcome}
\title{Creates long outcome table for fitting pooled logistic regression.}
\usage{
convertToLongOutcome(time, status, freqTime = 1, linearEffect = FALSE)
}
\arguments{
\item{time}{Numeric: Observed event time}

\item{status}{Numeric: Observed event status}

\item{freqTime}{Integer: Coarsen observed time to freqTime-day intervals}

\item{linearEffect}{Logical: Generate linear time effects}
}
\value{
A long outcome table for fitting pooled logistic regression.
}
\description{
\code{convertToLongOutcome} creates a long outcome table for fitting pooled logistic regression.
}
\examples{
time <- c(5, 4, 3, 2, 2, 2, 1, 1)
status <- c(0, 1, 0, 0, 1, 1, 0, 1)
convertToLongOutcome(time, status)
}
