% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeEffects.R
\name{convertToTimeVaryingCoef}
\alias{convertToTimeVaryingCoef}
\title{Convert short sparse covariate table to long sparse covariate table for time-varying coefficients.}
\usage{
convertToTimeVaryingCoef(shortCov, longOut, timeVaryCoefId)
}
\arguments{
\item{shortCov}{A data frame containing the covariate with predefined columns (see below), output of \code{splitTime}.}

\item{longOut}{A data frame containing the outcomes with predefined columns (see below).}

\item{timeVaryCoefId}{Integer: A numeric identifier of a time-varying coefficient}
}
\value{
A long sparse covariate table for time-varying coefficients.
}
\description{
\code{convertToTimeVaryingCoef} split the analysis time into several intervals for time-varying coefficients
}
\details{

}
