% Generated by roxygen2 (4.0.2): do not edit by hand
\name{coverage}
\alias{coverage}
\title{Coverage}
\usage{
coverage(goldStandard, lowerBounds, upperBounds)
}
\arguments{
\item{goldStandard}{Numeric vector}

\item{lowerBounds}{Numeric vector. Lower bound of the confidence intervals}

\item{upperBounds}{Numeric vector. Upper bound of the confidence intervals}
}
\value{
The proportion of times \code{goldStandard} falls between \code{lowerBound} and \code{upperBound}
}
\description{
\code{coverage} computes the coverage on confidence intervals
}
\keyword{internal}

