% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abridge.R
\name{createAbridgePrior}
\alias{createAbridgePrior}
\title{Create an ABRIDGE Cyclops prior object}
\usage{
createAbridgePrior(penalty = "bic", exclude = c(), forceIntercept = FALSE)
}
\arguments{
\item{penalty}{Specifies the ABRIDGE penalty; possible values are `BIC` or `AIC` or a numeric value}

\item{exclude}{A vector of numbers or covariateId names to exclude from prior}

\item{forceIntercept}{Logical: Force intercept coefficient into prior}
}
\value{
An ABRIDGE Cyclops prior object of class inheriting from \code{"cyclopsAbridgePrior"}
and \code{"cyclopsPrior"} for use with \code{fitCyclopsModel}.
}
\description{
\code{createAbridgePrior} creates an ABRIDGE Cyclops prior object for use with \code{\link{fitCyclopsModel}}.
}
\examples{
prior <- createAbridgePrior(penalty = "bic")

}

