% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecialPriors.R
\name{createNonSeparablePrior}
\alias{createNonSeparablePrior}
\title{Create a Cyclops prior object that returns the MLE of non-separable coefficients}
\usage{
createNonSeparablePrior(maxIterations = 10, ...)
}
\arguments{
\item{maxIterations}{Numeric: maximum iterations to achieve convergence}

\item{...}{Additional argument(s) for \code{\link{fitCyclopsModel}}}
}
\value{
A Cyclops prior object of class inheriting from
\code{"cyclopsPrior"} for use with \code{fitCyclopsModel}.
}
\description{
\code{createNonSeparablePrior} creates a Cyclops prior object for use with \code{\link{fitCyclopsModel}}.
}
\examples{
prior <- createNonSeparablePrior()

}
