% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createSqlCcdData}
\alias{createSqlCcdData}
\title{Create an OHDSI CCD data object from SQL input}
\usage{
createSqlCcdData(modelType, control)
}
\arguments{
\item{modelType}{character string: Valid types are listed below.}

\item{control}{An OHDSI CCD fit control object (optional)}
}
\description{
\code{createSqlCcdData} creates an empty OHDSI CCD data object into which data can be appended in chunks.
}
\section{Models}{

Currently supported model types are:
\tabular{ll}{
	\verb{	"ls"} \tab Least squares \cr
	\verb{	"pr"} \tab Poisson regression \cr
	\verb{	"lr"} \tab Logistic regression \cr
	\verb{	"clr"} \tab Conditional logistic regression \cr
 \verb{	"cpr"} \tab Conditional Poisson regression \cr
	\verb{	"sccs"} \tab Self-controlled case series \cr
	\verb{	"cox"} \tab Cox proportional hazards regression \cr
}
}

