% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{createSqlCyclopsData}
\alias{createSqlCyclopsData}
\title{Create an Cyclops data object from SQL input}
\usage{
createSqlCyclopsData(modelType, control)
}
\arguments{
\item{modelType}{character string: Valid types are listed below.}

\item{control}{An Cyclops fit control object (optional)}
}
\description{
\code{createSqlCyclopsData} creates an empty Cyclops data object into which data can be appended in chunks.
}
\section{Models}{

Currently supported model types are:
\tabular{ll}{
	\verb{	"ls"} \tab Least squares \cr
	\verb{	"pr"} \tab Poisson regression \cr
	\verb{	"lr"} \tab Logistic regression \cr
	\verb{	"clr"} \tab Conditional logistic regression \cr
 \verb{	"cpr"} \tab Conditional Poisson regression \cr
	\verb{	"sccs"} \tab Self-controlled case series \cr
	\verb{	"cox"} \tab Cox proportional hazards regression \cr
}
}

