% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightBasedSearchControl.R
\name{createWeightBasedSearchControl}
\alias{createWeightBasedSearchControl}
\title{Create a Cyclops control object that supports in- / out-of-sample hyperparameter search using weights}
\usage{
createWeightBasedSearchControl(cvType = "auto", initialValue = 1, ...)
}
\arguments{
\item{cvType}{Must equal "auto"}

\item{initialValue}{Initial value for auto-search parameter}

\item{...}{Additional parameters passed through to \code{\link{createControl}}}
}
\value{
A Cyclops prior object of class inheriting from \code{"cyclopsControl"}
for use with \code{fitCyclopsModel}.
}
\description{
\code{createWeightBasedSearchControl} creates a Cyclops control object for use with \code{\link{fitCyclopsModel}}
that supports hyperparameter optimization through an auto-search where weight = 1 identifies in-sample observations
and weight = 0 identifies out-of-sample observations.
}
