% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclops.R
\name{cyclops}
\alias{cyclops}
\alias{.onUnload}
\title{Cyclops: Cyclic coordinate descent for logistic, Poisson and survival analysis}
\usage{
.onUnload(libpath)
}
\description{
The Cyclops package incorporates cyclic coordinate descent and
majorization-minimization approaches to fit a variety of regression models
found in large-scale observational healthcare data.  Implementations focus
on computational optimization and fine-scale parallelization to yield
efficient inference in massive datasets.
}
