% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cyclopsCoxMM.R
\name{cyclopsCoxMM}
\alias{cyclopsCoxMM}
\title{Simulates data to test cox MM}
\usage{
cyclopsCoxMM(N, p, c, e, intercept = TRUE, beta = NA,
  simulateTies = FALSE)
}
\arguments{
\item{N}{number of patients}

\item{p}{number of covariates}

\item{c}{censoring probability}

\item{e}{distance between Poisson regressions to stop}

\item{intercept}{whether to include intercept in model, default = TRUE}

\item{beta}{vector to generate data from}

\item{simulateTies}{whether to simulate ties}
}
\value{
Returns a lit with the following: \describe{
\item{seedCoefficients}{This is equal to input beta}
\item{goldCox}{Cox regression model from using survival package function coxph}
\item{cyclopsCox}{Cox regression model from using cyclops}
\item{goldPoisson}{Cox MM poisson model from using glm}
\item{cyclopsPoisson}{Cox MM poisson model from using cyclops} }
}
\description{
Simulates data to test the cox MM algorithm.
}

