% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{cyclopsLibraryFileName}
\alias{cyclopsLibraryFileName}
\title{Get full path to installed Cyclops library}
\usage{
cyclopsLibraryFileName()
}
\value{
String
}
\description{
\code{cyclopsLibraryFileName} returns the full path to the installed Cyclops .so/.dll file
}
