% Generated by roxygen2 (4.0.1): do not edit by hand
\name{finalizeSqlCcdData}
\alias{finalizeSqlCcdData}
\title{finalizeSqlCcdData}
\usage{
finalizeSqlCcdData(object, addIntercept = FALSE, useOffsetCovariate = NULL,
  offsetAlreadyOnLogScale = FALSE, sortCovariates = FALSE,
  makeCovariatesDense = NULL)
}
\arguments{
\item{object}{CCD data object}

\item{addIntercept}{Add an intercept covariate if one was not imported through SQL}

\item{useOffsetCovariate}{Specify is a covariate should be used as an offset (fixed coefficient = 1).
Set option to \code{"useTime"} to specify the time-to-event column,
otherwise include a single numeric or character covariate name.}

\item{logOffset}{Set to \code{TRUE} to indicate that offsets were log-transformed before importing into CCD data object.}

\item{sortCovariates}{Sort covariates in numeric-order with intercept first if it exists.}

\item{makeCovariatesDense}{List of numeric or character covariates names to densely represent in CCD data object.
For efficiency, we suggest making atleast the intercept dense.}
}
\description{
\code{finalizeSqlCcdData} finalizes a CCD data object
}

