% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{finalizeSqlCyclopsData}
\alias{finalizeSqlCyclopsData}
\title{finalizeSqlCyclopsData}
\usage{
finalizeSqlCyclopsData(object, addIntercept = FALSE,
  useOffsetCovariate = NULL, offsetAlreadyOnLogScale = FALSE,
  sortCovariates = FALSE, makeCovariatesDense = NULL)
}
\arguments{
\item{object}{Cyclops data object}

\item{addIntercept}{Add an intercept covariate if one was not imported through SQL}

\item{useOffsetCovariate}{Specify is a covariate should be used as an offset (fixed coefficient = 1).
Set option to \code{-1} to specify the time-to-event column,
otherwise include a single numeric or character covariate name.}

\item{offsetAlreadyOnLogScale}{Set to \code{TRUE} to indicate that offsets were log-transformed before importing into Cyclops data object.}

\item{sortCovariates}{Sort covariates in numeric-order with intercept first if it exists.}

\item{makeCovariatesDense}{List of numeric or character covariates names to densely represent in Cyclops data object.
For efficiency, we suggest making atleast the intercept dense.}
}
\description{
\code{finalizeSqlCyclopsData} finalizes a Cyclops data object
}
\keyword{internal}
