% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitCcdModel}
\alias{fitCcdModel}
\title{fitCcdModel}
\usage{
fitCcdModel(ccdData, tolerance = 1e-08, variance, returnEstimates = TRUE,
  forceColdStart = FALSE)
}
\arguments{
\item{formula}{An R formula}

\item{data}{}
}
\value{
A list that contains a CCD model fit object pointer and an operation duration
}
\description{
\code{fitCcdModel} fits a CCD model data object
}
\details{
This function performs numerical optimization to fit a CCD model data object.
}
\examples{
## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
ccdData <- createCcdDataFrame(counts ~ outcome + treatment, modelType="pr")
ccdFit <- fitCcdModel(ccdData)
}

