% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{fitCyclopsSimulation}
\alias{fitCyclopsSimulation}
\title{Fit simulated data}
\usage{
fitCyclopsSimulation(
  sim,
  useCyclops = TRUE,
  model = "logistic",
  coverage = TRUE,
  includePenalty = FALSE,
  computeDevice = "native"
)
}
\arguments{
\item{sim}{A simulated Cyclops dataset generated via \code{simulateCyclopsData}}

\item{useCyclops}{Logical: use Cyclops or a standard routine}

\item{model}{String: Fitted regression model type}

\item{coverage}{Logical: report coverage statistics}

\item{includePenalty}{Logical: include regularized regression penalty in computing profile likelihood based confidence intervals}

\item{computeDevice}{String: Name of compute device to employ; defaults to \code{"native"} C++ on CPU}
}
\description{
\code{fitCyclopsSimulation} fits simulated Cyclops data using Cyclops or a standard routine.
This function is useful for simulation studies comparing the performance of Cyclops when considering
large, sparse datasets.
}
