% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{getCrossValidationInfo}
\alias{getCrossValidationInfo}
\title{Get cross-validation information from a Cyclops model fit}
\usage{
getCrossValidationInfo(object)
}
\arguments{
\item{object}{A Cyclops model fit object}
}
\description{
{getCrossValidationInfo} returns the predicted optimal cross-validation point and ordinate
}
\keyword{internal}
