% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFloatingPointSize}
\alias{getFloatingPointSize}
\title{Get floating point size}
\usage{
getFloatingPointSize(object)
}
\arguments{
\item{object}{A Cyclops data object}
}
\description{
\code{getFloatingPointSize} returns the floating-point representation size in a Cyclops data object
}
