% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getNumberOfRows}
\alias{getNumberOfRows}
\title{Get total number of rows}
\usage{
getNumberOfRows(object)
}
\arguments{
\item{object}{An OHDSI CCD data object}
}
\description{
\code{getNumberOfRows} returns the total number of outcome rows in an OHDSI CCD data object
}

