% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{getUnivariableCorrelation}
\alias{getUnivariableCorrelation}
\title{Get univariable correlation}
\usage{
getUnivariableCorrelation(cyclopsData, covariates = NULL,
  threshold = 0)
}
\arguments{
\item{cyclopsData}{A Cyclops data object}

\item{covariates}{Integer or string vector: list of covariates to report; default (NULL) implies all covariates}

\item{threshold}{Correlation threshold for reporting}
}
\value{
A list of covariates whose absolute correlation with the outcome is greater than or equal to the threshold
}
\description{
\code{getUnivariableCorrelation} reports covariates that have high correlation with the outcome
}
