% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{getUnivariableSeparability}
\alias{getUnivariableSeparability}
\title{Get univariable linear separability}
\usage{
getUnivariableSeparability(cyclopsData, covariates = NULL)
}
\arguments{
\item{cyclopsData}{A Cyclops data object}

\item{covariates}{Integer or string vector: list of covariates to report; default (NULL) implies all covariates}
}
\value{
A list of covariates that are univariably separable with the outcome
}
\description{
\code{getUnivariableSeparability} reports covariates that are univariably separable with the outcome
}
