% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{gradient}
\alias{gradient}
\title{Extract gradient}
\usage{
gradient(object)
}
\arguments{
\item{object}{A Cyclops model fit object}
}
\description{
\code{gradient} returns the current gradient wrt the regression parameters of
the log-likelihood of the fit in a Cyclops model fit object
}
