% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{isValidModelType}
\alias{isValidModelType}
\title{isValidModelType}
\usage{
isValidModelType(modelType)
}
\arguments{
\item{modelType}{character string: Valid types are listed below.}
}
\value{
TRUE/FALSE
}
\description{
\code{isValidModelType} checks for a valid Cyclops model type
}
\section{Models}{

Currently supported model types are:
\tabular{ll}{
	\verb{	"ls"} \tab Least squares \cr
	\verb{	"pr"} \tab Poisson regression \cr
	\verb{	"lr"} \tab Logistic regression \cr
	\verb{	"clr"} \tab Conditional logistic regression \cr
 \verb{	"cpr"} \tab Conditional Poisson regression \cr
	\verb{	"sccs"} \tab Self-controlled case series \cr
	\verb{	"cox"} \tab Cox proportional hazards regression \cr
}
}

