% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logLik.ccdFit}
\alias{logLik.ccdFit}
\title{Extract log-likelihood}
\usage{
\method{logLik}{ccdFit}(object, ...)
}
\arguments{
\item{object}{An OHDSI CCD model fit object}

\item{...}{Additional arguments}
}
\description{
\code{logLik} returns the current log-likelihood of the fit in an OHDSI CCD model fit object
}

