% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logLik.cyclopsFit}
\alias{logLik.cyclopsFit}
\title{Extract log-likelihood}
\usage{
\method{logLik}{cyclopsFit}(object, ...)
}
\arguments{
\item{object}{An OHDSI Cyclops model fit object}

\item{...}{Additional arguments}
}
\description{
\code{logLik} returns the current log-likelihood of the fit in an OHDSI Cyclops model fit object
}

