% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{meanLinearPredictor}
\alias{meanLinearPredictor}
\title{Calculates xbar*beta}
\usage{
meanLinearPredictor(cyclopsFit)
}
\arguments{
\item{cyclopsFit}{A Cyclops model fit object}
}
\description{
\code{meanLinearPredictor} computes xbar*beta for model fit
}
