% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleData.R
\docType{data}
\name{oxford}
\alias{oxford}
\title{Oxford self-controlled case series data}
\format{
A data frame with 38 rows and 6 variables:
\describe{
\item{indiv}{patient identifier}
\item{event}{number of events in interval}
\item{interval}{interval length in days}
\item{agegr}{age group}
\item{exgr}{exposure group}
\item{loginterval}{log interval length}
...
}
}
\usage{
data(oxford)
}
\description{
A dataset containing the MMR vaccination / meningitis in Oxford example from
Farrington and Whitaker.  There are 10 patients comprising 38 unique exposure intervals.
}
