% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCyclopsSimulationFit}
\alias{plotCyclopsSimulationFit}
\title{Plot Cyclops simulation model fit}
\usage{
plotCyclopsSimulationFit(fit, goldStandard, label)
}
\arguments{
\item{fit}{A Cyclops simulation fit generated by \code{fitCyclopsSimulation}}

\item{goldStandard}{Numeric vector.  True relative risks.}

\item{label}{String. Name of estimate type.}
}
\description{
\code{plotCyclopsSimulationFit} generates a plot that compares \code{goldStandard} coefficients to their
Cyclops model \code{fit}.
}

