% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.ccdFit}
\alias{predict.ccdFit}
\title{Model predictions}
\usage{
\method{predict}{ccdFit}(object, ...)
}
\arguments{
\item{object}{An OHDSI CCD model fit object}

\item{...}{Additional arguments}
}
\description{
\code{predict.ccdFit} computes model response-scale predictive values for all data rows
}

