% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ccdData}
\alias{print.ccdData}
\title{Print an OHDSI CCD data model object}
\usage{
\method{print}{ccdData}(x, show.call = TRUE, ...)
}
\arguments{
\item{x}{An OHDSI CCD data model object}

\item{show.call}{Logical: display last call to construct the OHDSI CCD data model object}

\item{...}{Additional arguments}
}
\description{
\code{print.ccdData} displays information about an OHDSI CCD data model object
}

