% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ccdFit}
\alias{print.ccdFit}
\title{Print an OHDSI CCD model fit object}
\usage{
\method{print}{ccdFit}(x, show.call = TRUE, ...)
}
\arguments{
\item{x}{An OHDSI CCD model fit object}

\item{show.call}{Logical: display last call to update the OHDSI CCD model fit object}

\item{...}{Additional arguments}
}
\description{
\code{print.ccdFit} displays information about an OHDSI CCD model fit object
}

