% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{print.cyclopsData}
\alias{print.cyclopsData}
\title{Print an OHDSI Cyclops data model object}
\usage{
\method{print}{cyclopsData}(x, show.call = TRUE, ...)
}
\arguments{
\item{x}{An OHDSI Cyclops data model object}

\item{show.call}{Logical: display last call to construct the OHDSI Cyclops data model object}

\item{...}{Additional arguments}
}
\description{
\code{print.cyclopsData} displays information about an OHDSI Cyclops data model object
}

