% Generated by roxygen2 (4.0.1): do not edit by hand
\name{printCcdRowIds}
\alias{printCcdRowIds}
\title{Print row identifiers}
\usage{
printCcdRowIds(object)
}
\arguments{
\item{object}{An OHDSI CCD data object}
}
\description{
\code{printCcdRowIds} return the row identifiers in an OHDSI CCD data object
}

