% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printMatrixMarket}
\alias{printMatrixMarket}
\title{Print Cyclops data matrix to file}
\usage{
printMatrixMarket(object, file)
}
\arguments{
\item{object}{A Cyclops data object}

\item{file}{Filename}
}
\description{
\code{printMatrixMarket} prints the data matrix to a file
}
\keyword{internal}
