% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prior}
\alias{prior}
\title{prior}
\usage{
prior(priorType, variance = 1, exclude = c(), useCrossValidation = FALSE,
  forceIntercept = FALSE)
}
\arguments{
\item{priorType}{}

\item{variance}{}

\item{exclude}{}

\item{useCrossValidation}{}
}
\value{
A CCD prior object of class inheriting from \code{"ccdPrior"} for use with \code{fitCcdModel}.
}
\description{
\code{prior} builds a CCD prior object
}
\section{Prior types}{

}

