% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readCcdData}
\alias{readCcdData}
\title{readCcdData}
\usage{
readCcdData(fileName, modelType)
}
\arguments{
\item{modelType}{character string: Valid types are listed below.}

\item{fileName}{Name of text file to be read. If fileName does not contain an absolute path,
the name is relative to the current working directory, \code{\link{getwd}}.}

\item{modelTypeName}{character string: if non-empty, declares the specific model formatting to
                                                 read.  Valid types are listed below.}
}
\value{
Returns an \code{\link{environment}} of class inheriting from \code{"ccdData"} that
contains at least the following objects:
\tabular{ll}{
     \code{  ccdDataPtr} \tab    An Rcpp \code{externalptr} to the C++ object that holds the data.
                                 The object contents does not get written to \code{.Rdata}, so
                                 \code{ccdData} are not reusable across restarts of \code{R}. \cr
     \code{  ccdInterfacePtr} \tab   An Rcpp \code{externalptr} to the C++ object that fits the data.
                                     Again, this is not reusable across \code{R} invokations. \cr
     \code{  timeLoad} \tab Amount of time (in seconds) taken to load the data. \cr
 }

The generic functions \code{\link{print}} and \code{\link{summary}} can be used to
display various summaries of the data.
}
\description{
\code{readCcdData} reads a CCD-formatted text file
}
\details{
This function reads a CCD-formatted text file and returns a CCD data object. The first line of the
file may start with '\samp{#}', indicating that it contains header options.  Valid header options are:

 \tabular{ll}{
  \verb{	row_label}		\tab (assume file contains a numeric column of unique row identifiers) \cr
  \verb{	stratum_label}\tab (assume file contains a numeric column of stratum identifiers) \cr
  \verb{	weight}				\tab (assume file contains a column of row-specific model weights, currently unused) \cr
	\verb{	offset}				\tab (assume file contains a dense column of linear predictor offsets) \cr
	\verb{	bbr_outcome}	\tab (assume logistic outcomes are encoded -1/+1 following BBR) \cr
	\verb{	log_offset}		\tab (assume file contains a dense column of values x_i for which log(x_i) is the offset) \cr
	\verb{	add_intercept}\tab (automatically include an intercept column of all 1s for each entry) \cr
	\verb{	indicator_only}\tab(assume all covariates 0/1-valued and only covariate name is given) \cr
	\verb{	sparse}				\tab (force all BBR formatted covariates to be represented as sparse, instead of sparse-indicator, columns .. really only for debugging) \cr
	\verb{	dense}				\tab (force all BBR formatted covariates to be represented as dense columns .. really only for debugging) \cr
}

Successive lines of the file are white-space delimited and follow the format:

\preformatted{	[Row ID] {Stratum ID} [Weight] <Outcome> {Censored} {Offset} <BBR covariates>}

\itemize{
  	\item \verb{[optional]}
  	\item \verb{<required>}
  	\item \verb{{required or optional depending on model}}
 }

Bayesian binary regression (BBR) covariates are white-space delimited and generally in a sparse
\samp{<name>:<value>} format, where \samp{name} must (currently) be numeric and \samp{value} is non-zero.
If option \samp{indicator_only} is specified, then format is simply \samp{<name>}.
\samp{Row ID} and \samp{Stratum ID} must be numeric, and rows must be sorted such that equal \samp{Stratum ID}
are consecutive.
\samp{Stratum ID} is required for \samp{clr} and \samp{sccs} models.
\samp{Censored} is required for a \samp{cox} model.
\samp{Offset} is (currently) required for a \samp{sccs} model.
}
\section{Models}{

Currently supported model types are:
\tabular{ll}{
	\verb{	"ls"} \tab Least squares \cr
	\verb{	"pr"} \tab Poisson regression \cr
	\verb{	"lr"} \tab Logistic regression \cr
	\verb{	"clr"} \tab Conditional logistic regression \cr
 \verb{	"cpr"} \tab Conditional Poisson regression \cr
	\verb{	"sccs"} \tab Self-controlled case series \cr
	\verb{	"cox"} \tab Cox proportional hazards regression \cr
}
}
\examples{
dataPtr = readCcdData(system.file("extdata/infert_ccd.txt", package="CCD"), "clr")
}

