% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{residuals.cyclopsFit}
\alias{residuals.cyclopsFit}
\title{Model residuals}
\usage{
\method{residuals}{cyclopsFit}(object, parm, type = "schoenfeld", ...)
}
\arguments{
\item{object}{A Cyclops model fit object}

\item{parm}{A specification of which parameters require residuals,
either a vector of numbers or covariateId names}

\item{type}{Character string indicating the type of residual desires. Possible
values are "schoenfeld".}
}
\description{
\code{residuals.cyclopsFit} computes model residuals for Cox model-based Cyclops objects
}
