% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{runBootstrap}
\alias{runBootstrap}
\title{Run bootstrap for Cyclops model parameters}
\usage{
runBootstrap(object, replicates, ciCoverage = 0.95)
}
\arguments{
\item{object}{A fitted Cyclops model object}

\item{replicates}{Numeric: number of bootstrap samples}

\item{ciCoverage}{Numeric (0-1): coverage proportion for bootstrap confidence intervals}
}
\description{
Run bootstrap for Cyclops model parameters
}
