% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cyclopsCoxMM.R
\name{simulateCoxData}
\alias{simulateCoxData}
\title{Simulates cox data}
\usage{
simulateCoxData(N, p, c, simulateTies = TRUE)
}
\arguments{
\item{N}{number of patients}

\item{p}{number of covariates}

\item{c}{censoring probability}

\item{simulateTies}{simulate ties by rounding to 3 significant digits}
}
\value{
\describe{\item{covariates}{matrix of covariates}
\item{outcomes}{matrix of outcomes including time to event and event type}}
}
\description{
Simulates cox data that can include ties. Represented as a matrix
}

