% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulateCyclopsData}
\alias{simulateCyclopsData}
\title{Simulation Cyclops dataset}
\usage{
simulateCyclopsData(nstrata = 200, nrows = 10000, ncovars = 20,
  effectSizeSd = 1, eCovarsPerRow = ncovars/100, model = "survival")
}
\arguments{
\item{nstrata}{Numeric: Number of strata}

\item{nrows}{Numeric: Number of observation rows}

\item{ncovars}{Numeric: Number of covariates}

\item{effectSizeSd}{Numeric: Standard derivation of the simulated regression coefficients}

\item{eCovarsPerRow}{Number: Effective number of non-zero covariates per data row}

\item{model}{String: Simulation model. Choices are: \code{logistic}, \code{poisson} or \code{survival}}
}
\value{
A simulated data set
}
\description{
\code{simulateCyclopsData} generates a simulated large, sparse data set for use by \code{fitCyclopsSimulation}.
}

