% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeEffects.R
\name{splitTime}
\alias{splitTime}
\title{Split the analysis time into several intervals for time-varying coefficients.}
\usage{
splitTime(time, y, cut)
}
\arguments{
\item{time}{Numeric: Observed event time}

\item{cut}{Numeric: Time points to cut at}

\item{status}{Numeric: Observed event status}
}
\value{
A long outcome table for time-varying coefficients.
}
\description{
\code{splitTime} split the analysis time into several intervals for time-varying coefficients
}
\examples{
time <- c(4.0, 3.5, 3, 2.5, 2, 1.5, 1)
status <- c(1, 1, 0, 1, 1, 0, 1)
cut <- c(2)
splitTime(time, status, cut)
}
