% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.ccdData}
\alias{summary.ccdData}
\title{OHDSI CCD data object summary}
\usage{
\method{summary}{ccdData}(object, ...)
}
\arguments{
\item{object}{An OHDSI CCD data object}

\item{...}{Additional arguments}
}
\value{
Returns a \code{data.frame} that reports simply summarize statistics for each covariate in an OHDSI CCD data object.
}
\description{
\code{summary.ccdData} summarizes the data held in an OHDSI CCD data object.
}

