% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survfit.R
\name{survfit.cyclopsFit}
\alias{survfit.cyclopsFit}
\title{Calculate baseline hazard function}
\usage{
\method{survfit}{cyclopsFit}(formula)
}
\arguments{
\item{formula}{A Cyclops survival model fit object}
}
\value{
Baseline survival function, not adjusted for each covariate to have mean 0 (unlike coxph{survival})
}
\description{
\code{survfit.cyclopsFit} computes baseline hazard function
}

