% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residuals.R
\name{testProportionality}
\alias{testProportionality}
\title{Test hazard ratio proportionality assumption}
\usage{
testProportionality(object, parm = NULL, transformedTimes = NULL)
}
\arguments{
\item{object}{A Cyclops model fit object}

\item{parm}{A specification of which parameters require a proportionality test,
either a vector of numbers or covariateId names}

\item{transformedTimes}{Vector of transformed time. If NULL, then the default "identity"
transform of \code{cox.zph} is used}
}
\description{
\code{testProportionality} tests the hazard ratio proportionality assumption
of a Cyclops model fit object using a score test
}
