% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{univariableCorrelation}
\alias{univariableCorrelation}
\title{Check univariable correlation}
\usage{
univariableCorrelation(cyclopsData, covariates = NULL, threshold = 0.9)
}
\arguments{
\item{cyclopsData}{A Cyclops data object}

\item{covariates}{Integer or string vector: list of covariates to report; default (NULL) implies all covariates}

\item{threshold}{Correlation threshold for reporting}
}
\value{
A list of covariates whose absolute correlation with the outcome is greater than the threshold
}
\description{
\code{univariableCorrelation} reports covariates that have high correlation with the outcome
}

