% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{coef.cyclopsFit}
\alias{coef.cyclopsFit}
\title{Extract model coefficients}
\usage{
\method{coef}{cyclopsFit}(object, rescale = FALSE, ignoreConvergence = FALSE, ...)
}
\arguments{
\item{object}{Cyclops model fit object}

\item{rescale}{Boolean: rescale coefficients for unnormalized covariate values}

\item{ignoreConvergence}{Boolean: return coefficients even if fit object did not converge}

\item{...}{Other arguments}
}
\value{
Named numeric vector of model coefficients.
}
\description{
\code{coef.cyclopsFit} extracts model coefficients from an Cyclops model fit object
}
