% Generated by roxygen2 (4.0.1): do not edit by hand
\name{confint.ccdFit}
\alias{confint.ccdFit}
\title{confint.ccdFit}
\usage{
\method{confint}{ccdFit}(object, parm, level = 0.95, control,
  overrideNoRegularization = FALSE, ...)
}
\arguments{
\item{fitted}{}

\item{covariates}{}
}
\value{
A \code{data.frame} containing profile 95% confidence intervals
}
\description{
\code{confinit.ccdFit} profiles the data likelihood to construct 95\% confidence intervals
}

