% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{convertToCyclopsVariance}
\alias{convertToCyclopsVariance}
\title{Convert to Cyclops Prior Variance}
\usage{
convertToCyclopsVariance(lambda, nobs)
}
\arguments{
\item{lambda}{Regularization parameter from \code{glmnet}}

\item{nobs}{Number of observation rows in dataset}
}
\value{
Prior variance under a Laplace() prior
}
\description{
\code{convertToCyclopsVariance} converts the regularization parameter \code{lambda}
from \code{glmnet} into a prior variance.
}
\keyword{internal}
