% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{coverage}
\alias{coverage}
\title{Coverage}
\usage{
coverage(goldStandard, lowerBounds, upperBounds)
}
\arguments{
\item{goldStandard}{Numeric vector}

\item{lowerBounds}{Numeric vector. Lower bound of the confidence intervals}

\item{upperBounds}{Numeric vector. Upper bound of the confidence intervals}
}
\value{
The proportion of times \code{goldStandard} falls between \code{lowerBound} and \code{upperBound}
}
\description{
\code{coverage} computes the coverage on confidence intervals
}

