% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createCcdDataFrame}
\alias{createCcdDataFrame}
\title{createCcdDataFrame}
\usage{
createCcdDataFrame(formula, sparseFormula, indicatorFormula, data, subset,
  weights, offs = NULL, y = NULL, z = NULL, dx = NULL, sx = NULL,
  ix = NULL, model = FALSE, method = "ccd.fit", ...)
}
\arguments{
\item{formula}{An R formula}

\item{data}{}
}
\value{
A list that contains a CCD model data object pointer and an operation duration
}
\description{
\code{createCcdDataFrame} creates a CCD model data object from an R formula
}
\details{
This function creates a CCD model data object from an R formula and data.frame.
}
\examples{
## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
ccdModelData <- createCcdDataFrame(counts ~ outcome + treatment, model=TRUE)
ccdModel <- ccdInitializeModel(ccdModelData$ccdModelDataPtr, modelType="pr", computeMLE=TRUE)
ccdFit <- ccdFitModel(ccdModel$interface)
ccdLog <- ccdLogModel(ccdModel$interface)
as.data.frame(ccdLog$estimation)


ccdModelData2 <- createCcdDataFrame(counts ~ outcome, indicatorFormula = ~ treatment, model=TRUE)
ccdModel2 <- ccdInitializeModel(ccdModelData2$ccdModelDataPtr, modelType="pr", computeMLE=TRUE)
ccdFit2 <- ccdFitModel(ccdModel2$interface)
ccdLog2 <- ccdLogModel(ccdModel2$interface)
as.data.frame(ccdLog2$estimation)
}

