% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createPrior}
\alias{createPrior}
\title{prior}
\usage{
createPrior(priorType, variance = 1, exclude = c(), graph = NULL,
  useCrossValidation = FALSE, forceIntercept = FALSE)
}
\arguments{
\item{priorType}{Character: specifies prior distribution.  See below for options}

\item{variance}{Numeric: prior distribution variance}

\item{exclude}{A vector of numbers or covariateId names to exclude from prior}

\item{graph}{Child-to-parent mapping for a hierarchical prior}

\item{useCrossValidation}{Logical: Perform cross-validation to determine prior \code{variance}.}

\item{forceIntercept}{Logical: Force intercept coefficient into prior}
}
\value{
A Cyclops prior object of class inheriting from \code{"cyclopsPrior"} for use with \code{fitCyclopsModel}.
}
\description{
\code{prior} builds a Cyclops prior object
}
\section{Prior types}{


We specify all priors in terms of their variance parameters.
Similar fitting tools for regularized regression often parameterize the Laplace distribution
in terms of a rate \code{"lambda"} per observation.
See \code{"glmnet"}, for example.

variance = 2 * / (nobs * lambda)^2 or lambda = sqrt(2 / variance) / nobs
}

