% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclops.R
\docType{package}
\name{cyclops}
\alias{Cyclops}
\alias{Cyclops-package}
\alias{cyclops}
\title{Cyclops: Cyclic coordinate descent for logistic, Poisson and survival analysis}
\description{
The Cyclops package incorporates cyclic coordinate descent and
majorization-minimization approaches to fit a variety of regression models
found in large-scale observational healthcare data.  Implementations focus
on computational optimization and fine-scale parallelization to yield
efficient inference in massive datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ohdsi/cyclops}
  \item Report bugs at \url{https://github.com/ohdsi/cyclops/issues}
}

}
\author{
\strong{Maintainer}: Marc A. Suchard \email{msuchard@ucla.edu}

Authors:
\itemize{
  \item Martijn J. Schuemie
  \item Trevor R. Shaddox
  \item Yuxi Tian
  \item Jianxiao Yang
  \item Eric Kawaguchi
}

Other contributors:
\itemize{
  \item Sushil Mittal [contributor]
  \item Observational Health Data Sciences and Informatics [copyright holder]
  \item Marcus Geelnard (provided the TinyThread library) [copyright holder, contributor]
  \item Rutgers University (provided the HParSearch routine) [copyright holder, contributor]
  \item R Development Core Team (provided the ZeroIn routine) [copyright holder, contributor]
}

}
