% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitCcdModel}
\alias{fitCcdModel}
\title{fitCcdModel}
\usage{
fitCcdModel(ccdData, prior, control, forceColdStart = FALSE,
  returnEstimates = TRUE)
}
\arguments{
\item{ccdData}{An OHDSI data object}

\item{prior}{A prior object. More details are given below.}
}
\value{
A list that contains a CCD model fit object pointer and an operation duration
}
\description{
\code{fitCcdModel} fits a CCD model data object
}
\details{
This function performs numerical optimization to fit a CCD model data object.
}
\section{Prior}{

Currently supported prior types are:
\tabular{ll}{
	\verb{	"none"} \tab Useful for finding MLE \cr
	\verb{	"laplace"} \tab L_1 regularization \cr
}
}
\examples{
## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
ccdData <- createCcdDataFrame(counts ~ outcome + treatment, modelType = "pr")
ccdFit <- fitCcdModel(ccdData, prior = prior("none"))
coef(ccdFit)
confint(ccdFit, c("outcome2","treatment3"))
predict(ccdFit)
}

