% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCovariateIds}
\alias{getCovariateIds}
\title{Get covariate identifiers}
\usage{
getCovariateIds(object)
}
\arguments{
\item{object}{An OHDSI Cyclops data object}
}
\description{
\code{getCovariateIds} returns a vector of integer covariate identifiers in an OHDSI Cyclops data object
}

