% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCovariateIds}
\alias{getCovariateIds}
\title{Get covariate identifiers}
\usage{
getCovariateIds(object)
}
\arguments{
\item{object}{A Cyclops data object}
}
\description{
\code{getCovariateIds} returns a vector of integer covariate identifiers in a Cyclops data object
}

