% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCovariateTypes}
\alias{getCovariateTypes}
\title{Get covariate types}
\usage{
getCovariateTypes(object, covariateLabel)
}
\arguments{
\item{object}{An OHDSI Cyclops data object}

\item{covariateLabel}{Integer vector: covariate identifiers to return}
}
\description{
\code{getCovariateTypes} returns a vector covariate types in an OHDSI Cyclops data object
}

