% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getCovariateTypes}
\alias{getCovariateTypes}
\title{Get covariate types}
\usage{
getCovariateTypes(object, covariateLabel)
}
\arguments{
\item{object}{A Cyclops data object}

\item{covariateLabel}{Integer vector: covariate identifiers to return}
}
\description{
\code{getCovariateTypes} returns a vector covariate types in a Cyclops data object
}

