% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{getCyclopsPredictiveLogLikelihood}
\alias{getCyclopsPredictiveLogLikelihood}
\title{Compute predictive log-likelihood from a Cyclops model fit}
\usage{
getCyclopsPredictiveLogLikelihood(object, weights)
}
\arguments{
\item{object}{A Cyclops model fit object}

\item{weights}{Numeric vector: vector of 0/1 identifying subset (=1) of rows from \code{object} to use in computing the log-likelihood}
}
\value{
The predictive log-likelihood
}
\description{
\code{getCyclopsPredictiveLogLikelihood} returns the log-likelihood of a subset of the data in a Cyclops model fit object.
}
\keyword{internal}
