% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{getCyclopsProfileLogLikelihood}
\alias{getCyclopsProfileLogLikelihood}
\title{Profile likelihood for Cyclops model parameters}
\usage{
getCyclopsProfileLogLikelihood(object, parm, x, includePenalty = TRUE)
}
\arguments{
\item{object}{A fitted Cyclops model object}

\item{parm}{A specification of which parameter requires profiling,
either a vector of numbers of covariateId names}

\item{x}{A vector of values of the parameter}

\item{includePenalty}{Logical: Include regularized covariate penalty in profile}
}
\value{
A vector of the profile log likelihood evaluated at x
}
\description{
\code{getCyclopsProfileLogLikelihood} evaluates the profile likelihood at a grid of parameter values.
}
