% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHyperParameter}
\alias{getHyperParameter}
\title{Get hyperparameter}
\usage{
getHyperParameter(object)
}
\arguments{
\item{object}{An OHDSI Cyclops model fit object}
}
\description{
\code{getHyperParameter} returns the current hyper parameter in an OHDSI Cyclops model fit object
}

