% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNumberOfCovariates}
\alias{getNumberOfCovariates}
\title{Get total number of covariates}
\usage{
getNumberOfCovariates(object)
}
\arguments{
\item{object}{An OHDSI Cyclops data object}
}
\description{
\code{getNumberOfCovariates} returns the total number of covariates in an OHDSI Cyclops data object
}

