% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNumberOfCovariates}
\alias{getNumberOfCovariates}
\title{Get total number of covariates}
\usage{
getNumberOfCovariates(object)
}
\arguments{
\item{object}{A Cyclops data object}
}
\description{
\code{getNumberOfCovariates} returns the total number of covariates in a Cyclops data object
}
