% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNumberOfRows}
\alias{getNumberOfRows}
\title{Get total number of rows}
\usage{
getNumberOfRows(object)
}
\arguments{
\item{object}{An OHDSI Cyclops data object}
}
\description{
\code{getNumberOfRows} returns the total number of outcome rows in an OHDSI Cyclops data object
}

