% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNumberOfRows}
\alias{getNumberOfRows}
\title{Get total number of rows}
\usage{
getNumberOfRows(object)
}
\arguments{
\item{object}{A Cyclops data object}
}
\description{
\code{getNumberOfRows} returns the total number of outcome rows in a Cyclops data object
}

