% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNumberOfStrata}
\alias{getNumberOfStrata}
\title{Get number of strata}
\usage{
getNumberOfStrata(object)
}
\arguments{
\item{object}{An OHDSI Cyclops data object}
}
\description{
\code{getNumberOfStrata} return the number of unique strata in an OHDSI Cyclops data object
}

