% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNumberOfTypes}
\alias{getNumberOfTypes}
\title{Get total number of outcome types}
\usage{
getNumberOfTypes(object)
}
\arguments{
\item{object}{An OHDSI Cyclops data object}
}
\description{
\code{getNumberOfTypes} returns the total number of outcome types in an OHDSI Cyclops data object
}

