% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getNumberOfTypes}
\alias{getNumberOfTypes}
\title{Get total number of outcome types}
\usage{
getNumberOfTypes(object)
}
\arguments{
\item{object}{A Cyclops data object}
}
\description{
\code{getNumberOfTypes} returns the total number of outcome types in a Cyclops data object
}
\keyword{internal}

