% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{mse}
\alias{mse}
\title{Mean squared error}
\usage{
mse(goldStandard, estimates)
}
\arguments{
\item{goldStandard}{Numeric vector}

\item{estimates}{Numeric vector}
}
\value{
MSE(\code{goldStandard}, \code{estimates})
}
\description{
\code{mse} computes the mean squared error between two numeric vectors
}

