% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.cyclopsFit}
\alias{predict.cyclopsFit}
\title{Model predictions}
\usage{
\method{predict}{cyclopsFit}(object, ...)
}
\arguments{
\item{object}{A Cyclops model fit object}

\item{...}{Additional arguments}
}
\description{
\code{predict.cyclopsFit} computes model response-scale predictive values for all data rows
}

