% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict.cyclopsFit}
\alias{predict.cyclopsFit}
\title{Model predictions}
\usage{
\method{predict}{cyclopsFit}(object, newOutcomes, newCovariates, ...)
}
\arguments{
\item{object}{A Cyclops model fit object}

\item{newOutcomes}{An optional data frame or ffdf object, similar to the object used in \code{\link{convertToCyclopsData}}.}

\item{newCovariates}{An optional data frame or ffdf object, similar to the object used in \code{\link{convertToCyclopsData}}.}

\item{...}{Additional arguments}
}
\description{
\code{predict.cyclopsFit} computes model response-scale predictive values for all data rows
}

