% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{print.cyclopsData}
\alias{print.cyclopsData}
\title{Print a Cyclops data object}
\usage{
\method{print}{cyclopsData}(x, show.call = TRUE, ...)
}
\arguments{
\item{x}{A Cyclops data model object}

\item{show.call}{Logical: display last call to construct the Cyclops data model object}

\item{...}{Additional arguments}
}
\description{
\code{print.cyclopsData} displays information about a Cyclops data model object.
}

