% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.cyclopsFit}
\alias{print.cyclopsFit}
\title{Print a Cyclops model fit object}
\usage{
\method{print}{cyclopsFit}(x, show.call = TRUE, ...)
}
\arguments{
\item{x}{A Cyclops model fit object}

\item{show.call}{Logical: display last call to update the Cyclops model fit object}

\item{...}{Additional arguments}
}
\description{
\code{print.cyclopsFit} displays information about a Cyclops model fit object
}

