% Generated by roxygen2 (4.0.2): do not edit by hand
\name{printCyclopsRowIds}
\alias{printCyclopsRowIds}
\title{Print row identifiers}
\usage{
printCyclopsRowIds(object)
}
\arguments{
\item{object}{An OHDSI Cyclops data object}
}
\description{
\code{printCcdRowIds} return the row identifiers in an OHDSI Cyclops data object
}

