% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{printCyclopsRowIds}
\alias{printCyclopsRowIds}
\title{Print row identifiers}
\usage{
printCyclopsRowIds(object)
}
\arguments{
\item{object}{A Cyclops data object}
}
\description{
\code{printCcdRowIds} return the row identifiers in a Cyclops data object
}
\keyword{internal}

