% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{runBootstrap}
\alias{runBootstrap}
\title{Run Bootstrap for Cyclops model parameter}
\usage{
runBootstrap(object, outFileName, treatmentId, replicates)
}
\arguments{
\item{object}{A fitted Cyclops model object}

\item{outFileName}{Character: Output file name}

\item{treatmentId}{Character: variable to output}

\item{replicates}{Numeric: number of bootstrap samples}
}
\description{
Run Bootstrap for Cyclops model parameter
}
