% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeEffects.R
\name{splitTime}
\alias{splitTime}
\title{Split the analysis time into several intervals for time-varying coefficients.}
\usage{
splitTime(shortOut, cut)
}
\arguments{
\item{shortOut}{A data frame containing the outcomes with predefined columns (see below).}

\item{cut}{Numeric: Time points to cut at

These columns are expected in the shortOut object:
\tabular{lll}{
\verb{rowId}   \tab(integer) \tab Row ID is used to link multiple covariates (x) to a single outcome (y) \cr
\verb{y}       \tab(real) \tab Observed event status \cr
\verb{time}    \tab(real) \tab Observed event time \cr
}}
}
\value{
A long outcome table for time-varying coefficients.
}
\description{
\code{splitTime} split the analysis time into several intervals for time-varying coefficients
}
