% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residuals.R
\name{testProportionality}
\alias{testProportionality}
\title{Test hazard ratio proportionality assumption}
\usage{
testProportionality(object, parm = NULL, transform = "km")
}
\arguments{
\item{object}{A Cyclops model fit object}

\item{parm}{A specification of which parameters require a proportionality test,
either a vector of numbers or covariateId names}

\item{transform}{Same as \code{transform} in \code{cox.zph}}
}
\description{
\code{testProportionality} tests the hazard ratio proportionality assumption
of a Cyclops model fit object using a score test
}
