% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelFit.R
\name{vcov.cyclopsFit}
\alias{vcov.cyclopsFit}
\title{Calculate variance-covariance matrix for a fitted Cyclops model object}
\usage{
\method{vcov}{cyclopsFit}(object, control, overrideNoRegularization = FALSE,
  ...)
}
\arguments{
\item{object}{A fitted Cyclops model object}

\item{control}{A Cyclops \code{\link{control}} object}

\item{overrideNoRegularization}{Logical: Enable variance-covariance estimation for regularized parameters}

\item{...}{Additional argument(s) for methods}
}
\value{
A matrix of the estimates covariances between all covariate estimates.
}
\description{
\code{vcov.cyclopsFit} returns the variance-covariance matrix for all covariates of a Cyclops model object
}

